Support non-rdf tests
