+++ src/main/java/org/w3/ldp/testsuite/test/BasicContainerTest.java
@@ -1,6 +1,7 @@
 package org.w3.ldp.testsuite.test;
 
 import com.jayway.restassured.response.Response;
+import org.testng.annotations.AfterSuite;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Optional;
 import org.testng.annotations.Parameters;
@@ -88,4 +89,17 @@ public class BasicContainerTest extends CommonContainerTest {
 		return basicContainer;
 	}
 
+	@AfterSuite(alwaysRun = true)
+	public void commonTearDown() {
+		if (httpLog != null) {
+			httpLog.println();
+			httpLog.flush();
+			httpLog.close();
+		}
+		if (skipLog != null) {
+			skipLog.println();
+			skipLog.flush();
+			skipLog.close();
+		}
+	}
 }
