+++ src/main/java/org/w3/ldp/testsuite/test/DirectContainerTest.java
@@ -5,6 +5,7 @@ import org.apache.jena.rdf.model.Resource;
 import com.jayway.restassured.response.Header;
 import com.jayway.restassured.response.Response;
 import org.apache.http.HttpStatus;
+import org.testng.annotations.AfterSuite;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Optional;
 import org.testng.annotations.Parameters;
@@ -418,4 +419,17 @@ public class DirectContainerTest extends CommonContainerTest {
 		return directContainer;
 	}
 
+	@AfterSuite(alwaysRun = true)
+	public void commonTearDown() {
+		if (httpLog != null) {
+			httpLog.println();
+			httpLog.flush();
+			httpLog.close();
+		}
+		if (skipLog != null) {
+			skipLog.println();
+			skipLog.flush();
+			skipLog.close();
+		}
+	}
 }
